Moved operations to separate files
