+++ src/main/index.ts
@@ -1,1 +1,6 @@
-export * from './trie-utils';
+export * from './setTrie';
+export * from './createTrie';
+export * from './setAllTrie';
+export * from './getTrie';
+export * from './searchTrie';
+export * from './trie-types';
